+++ other/AutomaticPause/source/AutomaticPauseManager/AutomaticPauseManager.c
@@ -107,7 +107,7 @@ void AutomaticPauseManager::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool AutomaticPauseManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool AutomaticPauseManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
