+++ other/LowPowerIndicator/source/LowPowerManager/LowPowerManager.c
@@ -54,7 +54,7 @@ void LowPowerManager::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool LowPowerManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool LowPowerManager::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
